Use `appTag` label filter for `getAppFromTag` method
The current approach gets all the apps from K8s and then filter by
`appTag` locally to get the target apps. This will put much pressure on
K8s API server when there are lots of Spark apps.

This commit overrides the `getApplications` method to support `appTag`
label as filter. With the new method it will be a server side filter and
K8s API will only return apps with the target `appTag`.