Add new configuration livy.server.recovery.persist-success-sessions
The new configuration is used to decide whether to persist successful sessions and recover them when Livy restarts.

Note: In Kubernetes environments, when there are thousands of successful sessions to recover, the Kube API could be
exhausted, potentially leading to performance issues or failures during recovery. Consider setting this to false
under such circumstances.

Default value is true.

If set to false, all successful sessions will be lost when Livy restarts.