Fix `RSCConf#isRunningOnKubernetes` logic for RSCDriver
The method `RSCConf#isRunningOnKubernetes` always returns false, since
`livy.spark.master` will not be get by `RSCConf`. The only useage of
this method is for setting `RPC_SERVER_ADDRESS` in `RSCDriver` when
Spark's running on Kubernetes, and it's causing Spark driver to fail
in interactive mode.

Let's remove this bug method, and use Spark conf directly to check if
it's running on Kubernetes.